% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsMBRM.R
\name{residuals.MRM}
\alias{residuals.MRM}
\title{Compute the randomized quantile residuals for MRM model}
\usage{
\method{residuals}{MRM}(object)
}
\arguments{
\item{object}{An object of class \code{MRM}, typically returned from \code{\link{MRMfit}}.}
}
\value{
A numeric vector of randomized quantile residuals.
}
\description{
This function computes the randomized quantile residuals for objects of class \code{MRM}.
}
\examples{
data(Arthritis1)
fit <- MRMfit(y ~ Sex + Age + Treatment + Time, data = Arthritis1)
summary(residuals(fit))

}
