% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsMBRM.R
\name{residuals.MRM}
\alias{residuals.MRM}
\title{Compute randomized quantile residuals for MRM model}
\usage{
\method{residuals}{MRM}(
  object,
  envelope = FALSE,
  R = 100,
  random = "gengamma",
  link = "cloglog",
  ...
)
}
\arguments{
\item{object}{An object of class \code{MRM}, typically returned from \code{\link{MRMfit}}.}

\item{envelope}{Logical. If \code{TRUE}, generates an envelope QQ-plot using Monte Carlo simulations.}

\item{R}{Integer. Number of replications to simulate the envelope (default is 100).}

\item{random}{Character. The distribution for the random effects. Possible values are \code{"normal"} and \code{"gengamma"}.}

\item{link}{Character. The link function used in the model. Examples include \code{"cloglog"} and \code{"logit"}.}

\item{...}{Additional arguments passed to \code{\link{MRMfit}} when \code{envelope = TRUE}.}
}
\value{
A numeric vector of randomized quantile residuals, or a QQ-plot with envelope if \code{envelope = TRUE}.
}
\description{
This function computes randomized quantile residuals for objects of class \code{MRM}.
}
\details{
If \code{envelope = TRUE}, the function also performs a Monte Carlo simulation to generate a normal QQ-plot with an envelope.
}
\examples{
data(Arthritis1)
fit <- MRMfit(y ~ Sex + Age + Treatment + Time, data = Arthritis1)
summary(residuals(fit))

}
